feat: Collapse db and expiration by default
https://harperdb.atlassian.net/browse/STUDIO-463
