+++ src/integrations/api/instance/database/tableNameSchema.tsx
@@ -1,5 +1,5 @@
-import { schemaRegex } from '@/features/instance/databases/modals/schemaRegex';
 import { z } from 'zod';
+import { schemaRegex } from './schemaRegex';
 
 export const tableNameSchema = z
 	.string()
